Create day1.py
Hello, I worked for roughly 80 minutes on the first three questions for Project Euler. Starting with question 1, I struggled the least and found thhat the easiest way to the solution was to begin with a variable. For question2, I  ran into quite a few roadblocks. Specifically, I tried to write a code that would complete the fibonacci sequence quickly but found out upon running it that too many errors came up for it to be efficient. I moved on and tried the establishing variables method again which worked well. For question 3, I initially thhought that running while loops was a smart way to approach it since the number given in the problem was so large. This ended up working very well for me and I worked through it by creating variables that would allow me to manipulate the number into finding the largest prime factor.